# src/server/api/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Acontext.",
        "title": "Acontext API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/project": {
            "post": {
                "security": [
                    {
                        "RootAuth": []
                    }
                ],
                "description": "Create a new project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Create project",
                "parameters": [
                    {
                        "description": "CreateProject payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateProjectReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Project"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/space": {
            "post": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Create a new space under a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Create space",
                "parameters": [
                    {
                        "description": "CreateSpace payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateSpaceReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Space"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/space/{space_id}": {
            "delete": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Delete a space by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Delete space",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/space/{space_id}/configs": {
            "get": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Retrieve the configurations of a space by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Get space configs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Space"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Update the configurations of a space by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Update space configs",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "UpdateConfigs payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateSpaceConfigsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/space/{space_id}/semantic_answer": {
            "get": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Retrieve the semantic answer for a given query within a space by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Get semantic answer",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GetSemanticAnswer payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GetSemanticAnswerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/space/{space_id}/semantic_global": {
            "get": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Retrieve the semantic global information for a given query within a space by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Get semantic global",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GetSemanticGlobal payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GetSemanticGlobalReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/space/{space_id}/semantic_grep": {
            "get": {
                "security": [
                    {
                        "ProjectAuth": []
                    }
                ],
                "description": "Retrieve the semantic grep results for a given query within a space by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "space"
                ],
                "summary": "Get semantic grep",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000",
                        "description": "Space ID",
                        "name": "space_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "GetSemanticGrep payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GetSemanticGrepReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.CreateProjectReq": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "handler.CreateSpaceReq": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "handler.GetSemanticAnswerReq": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "handler.GetSemanticGlobalReq": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "handler.GetSemanticGrepReq": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
...
        "handler.UpdateSpaceConfigsReq": {
            "type": "object",
            "required": [
                "configs"
            ],
            "properties": {
                "configs": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "model.Asset": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_seconds": {
                    "type": "number"
                },
                "etag": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "description": "Asset \u003c-\u003e Message",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "mime": {
                    "type": "string"
                },
                "s3_key": {
                    "type": "string"
                },
                "sha256": {
                    "type": "string"
                },
                "size_b": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "assets": {
                    "description": "Message \u003c-\u003e Asset",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Asset"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message_assets": {
                    "description": "Message \u003c-\u003e MessageAsset",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MessageAsset"
                    }
                },
                "parts": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "role": {
                    "type": "string"
                },
                "session": {
                    "description": "Message \u003c-\u003e Session",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Session"
                        }
                    ]
                },
                "session_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.MessageAsset": {
            "type": "object",
            "properties": {
                "asset": {
                    "description": "MessageAsset \u003c-\u003e Asset",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Asset"
                        }
                    ]
                },
                "asset_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "message": {
                    "description": "MessageAsset \u003c-\u003e Message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Message"
                        }
                    ]
                },
                "message_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Project": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "sessions": {
                    "description": "Project \u003c-\u003e Session",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Session"
                    }
                },
                "spaces": {
                    "description": "Project \u003c-\u003e Space",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Session": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "description": "Session \u003c-\u003e Message",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "project": {
                    "description": "Session \u003c-\u003e Project",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Project"
                        }
                    ]
                },
                "project_id": {
                    "type": "string"
                },
                "space": {
                    "description": "Session \u003c-\u003e Space",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Space"
                        }
                    ]
                },
                "space_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Space": {
            "type": "object",
            "properties": {
                "configs": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "project": {
                    "description": "Space \u003c-\u003e Project",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Project"
                        }
                    ]
                },
                "project_id": {
                    "type": "string"
                },
                "sessions": {
                    "description": "Space \u003c-\u003e Session",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Session"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ProjectAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RootAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}